fix(DataTable): move the component into Organisms
