+++ src/index.ts
@@ -70,11 +70,11 @@ export { default as NumberField, INumberFieldProps } from "./components/molecule
 export { default as InteractiveCard, IInteractiveCardProps } from "./components/molecules/InteractiveCard";
 export { default as Section, ISectionProps } from "./components/molecules/Section";
 export { default as Image, IImageProps } from "./components/molecules/Image";
-export { default as DataTable, IDataTableProps } from "./components/molecules/DataTable";
 export { default as OTPField, IOTPFieldProps } from "./components/molecules/OTPField";
 export { default as Breadcrumb, IBreadcrumbProps, IBreadcrumbItemProps } from "./components/molecules/Breadcrumb";
 export { default as PageHeader, IPageHeaderProps } from "./components/molecules/PageHeader";
 // Organisms
+export { default as DataTable, IDataTableProps } from "@components/organisms/DataTable";
 export { default as GlobalHeader, IGlobalHeaderProps, IAction, IProducts } from "./components/organisms/GlobalHeader";
 export { default as DataCardList, IDataCardListProps } from "@components/organisms/DataCardList";
